scrapers/findartinfo/requirements.txt
requests
bs4
lxml
django
tqdm
python-telegram-bot
django-dotenv